fix sonar error
